+++ bin/store/migrations/2025-08-06-000427_add_account_profile_fields/down.sql
@@ -1,0 +1,6 @@
+-- This file should undo anything in `up.sql`
+
+ALTER TABLE "account_profiles" 
+DROP COLUMN "date_of_birth",
+DROP COLUMN "middle_name",
+DROP COLUMN "auth_preference";